// web-client/src/public/components/Tasks/types.ts
/* eslint-disable */
/* prettier-ignore */
export enum ETaskListStatus {
  WaitingForAction = 'waiting-for-action',
  Loading = 'loading',
  Searching = 'searching',
  NoTasks = 'no-tasks',
  EmptySearchResult = 'empty-search-result',
  LastTaskFinished = 'last-task-finished',
  LastFilteredTaskFinished = 'last-filtered-task-finished',
}